Adjust card design
